split train process service
